test(srt): split seat helper regression coverage
